requirements.txt
numpy #==1.22.4
pandas #==1.4.2
scikit-learn #==1.1.1
streamlit

xgboost
scikeras
catboost

pytest
pytest-cov
tabulate

# Temporarily remove for heroku slug
#jupyter #==1.0.0
#matplotlib #==3.5.2
#matplotlib-inline #==0.1.3
#notebook #==6.4.12
#seaborn #==0.11.2

#xxxtensorflow
#tensorflow-cpu

#html2text #==2020.1.16
#ipykernel #==6.13.1
#regex #==2022.7.9
#requests #==2.28.0
#scipy #==1.8.1
